create new Methods class
